+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+
+## [15.1.0][15.1.0] - 2016-08-30
 ### Added
 - [Feature] Add a `verticalAlign` prop to `Table.TD` component.
 - [Feature] Add the ability to pass a `width` to `Table.TD` and `Table.TH` components.
@@ -502,7 +504,7 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 ### Added
 - [Patch] Use semantic versioning in LEGO (#58).
 
-[Unreleased]: https://github.com/optimizely/oui/compare/v15.0.0...HEAD
+[Unreleased]: https://github.com/optimizely/oui/compare/v15.1.0...HEAD
 [0.0.2]: https://github.com/optimizely/oui/compare/v0.0.1...v0.0.2
 [0.0.3]: https://github.com/optimizely/oui/compare/v0.0.2...v0.0.3
 [1.0.0]: https://github.com/optimizely/oui/compare/v0.0.3...v1.0.0
@@ -548,3 +550,4 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 [14.0.5]: https://github.com/optimizely/oui/compare/v14.0.0...v14.0.5
 [14.1.0]: https://github.com/optimizely/oui/compare/v14.0.5...v14.1.0
 [15.0.0]: https://github.com/optimizely/oui/compare/v14.1.0...v15.0.0
+[15.1.0]: https://github.com/optimizely/oui/compare/v15.0.0...v15.1.0